backend/requirements.txt
flask
flask-cors
pymongo
certifi
tqdm
python-dotenv
google-genai
requests
